requirements.txt
streamlit>=1.30.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
phonenumbers>=8.13.0
openai>=1.0.0
crawl4ai>=0.7.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
mistralai>=1.0.0
